Support mapping Ctrl to Esc on Release
Adds option to make ctrl act as esc if released without pressing any
other keys. (If other keys are pressed, it'll continue to function as a
regular ctrl modifier.)
This supports mapping caps-lock to ctrl (either via iVim or system
settings). So if this new setting is turned on, both caps-lock
can also function as both ctrl and esc.

note: only tested on iPad and in combination with system-settings and
iVim's caps-lock settings.

Tests:
Editted text while randomly pressing and holding physical caps-lock or ctrl keys.
- When neither iVim nor System is changing Caps-Lock or Ctrl behavior
- When iVim's Caps-Lock is mapped to Ctrl
- When iVim's Caps-Lock is mapped to Esc
- When System's Caps-Lock is mapped to Ctrl
- When iVim's Caps-Lock is mapped to Ctrl and Ctrl-to-Esc-on-release is
  turned on
- When System's Caps-Lock is mapped to Ctrl and iVim's
  Ctrl-to-Esc-on-release setting is turned on
- When both iVim and System are mapping Caps-Lock to Ctrl and iVim's
  Ctrl-to-Esc-on-release setting is turned on